create a PatternList, organize sizing/resolution
