+++ modules/common/utils/size.ts
@@ -1,0 +1,3 @@
+import { Vector2 } from "three";
+
+export const LED_COUNTS = new Vector2(96, 75);
