+++ modules/components/BlockView.tsx
@@ -1,11 +1,11 @@
 import { Block } from "../common/types/Block";
-import { ShaderMaterial } from "three";
+import { ShaderMaterial, Vector2 } from "three";
 import { StandardParams } from "../common/types/PatternParams";
-import { useFrame } from "@react-three/fiber";
-import { useRef } from "react";
-import vert from "@/modules/patterns/shaders/default.vert";
+import { useFrame, useThree } from "@react-three/fiber";
+import { useEffect, useRef } from "react";
 import { observer } from "mobx-react-lite";
 import { useStore } from "@/modules/common/types/StoreContext";
+import vert from "@/modules/patterns/shaders/default.vert";
 
 type BlockViewProps = {
   autorun?: boolean;
@@ -13,10 +13,20 @@ type BlockViewProps = {
 };
 
 export default observer(function BlockView({ autorun, block }: BlockViewProps) {
-  const { globalTime } = useStore();
   const shaderMaterial = useRef<ShaderMaterial>(null);
 
+  const { size } = useThree();
+  useEffect(() => {
+    // whenever the block or canvas size changes, reset the shader resolution
+    block.pattern.paramValues.u_resolution = new Vector2(
+      size.width,
+      size.height,
+    );
+  }, [block, size.width, size.height]);
+
+  // dereference block.startTime here so that we are not accessing an observable inside the useFrame callback
   const { startTime } = block;
+  const { globalTime } = useStore();
   useFrame(({ clock }) => {
     if (autorun) {
       block.update(clock.elapsedTime, clock.elapsedTime);
