+++ modules/components/Display.tsx
@@ -3,9 +3,10 @@ import { Box, HStack } from "@chakra-ui/react";
 import { Canvas } from "@react-three/fiber";
 import { observer } from "mobx-react-lite";
 import { useStore } from "@/modules/common/types/StoreContext";
+import { LED_COUNTS } from "@/modules/common/utils/size";
 
+// Multiplier on the base LED size for the main display only
 const DISPLAY_FACTOR = 3;
-const LED_COUNTS = { width: 96, height: 75 };
 
 type DisplayProps = {};
 
@@ -14,8 +15,8 @@ export default observer(function Display({}: DisplayProps) {
   return (
     <HStack py={4} justify="center">
       <Box
-        width={`${LED_COUNTS.width * DISPLAY_FACTOR}px`}
-        height={`${LED_COUNTS.height * DISPLAY_FACTOR}px`}
+        width={`${LED_COUNTS.x * DISPLAY_FACTOR}px`}
+        height={`${LED_COUNTS.y * DISPLAY_FACTOR}px`}
       >
         <Canvas>
           {currentBlock && (
