+++ modules/components/Editor.tsx
@@ -1,23 +1,15 @@
 import Timeline from "@/modules/components/Timeline";
-import { patterns } from "@/modules/patterns/Patterns";
-import { Box, Button, Grid, GridItem, Heading, VStack } from "@chakra-ui/react";
-import { useMemo, useState } from "react";
-import { Block } from "../common/types/Block";
+import { Box, Grid, GridItem, Heading, VStack } from "@chakra-ui/react";
 import Display from "@/modules/components/Display";
-import BlockView from "@/modules/components/BlockView";
-import { Canvas } from "@react-three/fiber";
+import PatternList from "@/modules/components/PatternList";
 
 export default function Editor() {
-  const [pattern, setPattern] = useState(patterns[0]);
-
-  const block = useMemo(() => new Block(pattern), [pattern]);
-
   return (
     <Box w="100vw" h="100vh">
       <Grid
-        templateAreas={`".      header"
-                        "nav    display"
-                        "nav    timeline"`}
+        templateAreas={`".        header"
+                        "patterns display"
+                        "patterns timeline"`}
         gridTemplateColumns="150px 1fr"
         gap="1"
         fontWeight="bold"
@@ -27,22 +19,8 @@ export default function Editor() {
             <Heading>Conjurer</Heading>
           </VStack>
         </GridItem>
-        <GridItem pl="2" area="nav">
-          <VStack>
-            <Canvas>
-              <BlockView key={block.pattern.name} autorun block={block} />
-            </Canvas>
-            {patterns.map((p) => (
-              <Button
-                key={p.name}
-                colorScheme="teal"
-                variant="outline"
-                onClick={() => setPattern(p)}
-              >
-                {p.name}
-              </Button>
-            ))}
-          </VStack>
+        <GridItem pl="2" area="patterns">
+          <PatternList />
         </GridItem>
         <GridItem pl="2" area="display">
           <Display />
